Integrate terrain rendering, waterways, and free camera mechanics into the main game
- Add `TerrainRenderer` to render real elevation data (`heightmap.json`) and waterways (`waterways.json`) using OpenStreetMap and EU-DEM sources.
- Replace static camera with a shared `FreeCameraRig` supporting pan, zoom, and rotate functionality.
- Update `LoadedMap` to replace terrain size with `CampCenter`, ensuring correct ground placement for entities.
- Adjust `MapLoader` and tests to accommodate terrain-based positions.
- Enhance `WorldPresenter` to place all entities on terrain height.
- Extend `generate_sprites.py` with conifer tree and rock pile sprites, adding static backdrops to the environment.
- Update resource file imports (`conifer_tree.png`, `rock_pile.png`) and Godot project configuration for terrain assets.

+++ src/ManyWinters.Core/Persistence/SaveData.cs
@@ -20,8 +20,8 @@ public sealed record SaveData(
 public sealed record PersonSaveData(
     int Id,
     string Name,
-    float PositionX,
-    float PositionY,
+    double PositionX,
+    double PositionY,
     bool IsAlive,
     float Hunger,
     float Fatigue,
@@ -39,23 +39,23 @@ public sealed record ItemStackSaveData(ItemKindId Kind, int Count);
 public sealed record ResourceNodeSaveData(
     int Id,
     ResourceKindId Kind,
-    float PositionX,
-    float PositionY,
+    double PositionX,
+    double PositionY,
     float RemainingAmount,
     float MaxAmount);
 
 public sealed record BuildingSaveData(
     int Id,
     BuildingKindId Kind,
-    float PositionX,
-    float PositionY,
+    double PositionX,
+    double PositionY,
     float Condition,
     IReadOnlyList<ItemStackSaveData> Inventory);
 
 public sealed record GraveSaveData(
     int Id,
-    float PositionX,
-    float PositionY,
+    double PositionX,
+    double PositionY,
     bool IsMarked,
     string? Name,
     int? AgeAtDeath,
