+++ src/ManyWinters.Core/World/Position.cs
@@ -1,3 +1,3 @@
 namespace ManyWinters.Core.World;
 
-public readonly record struct Position(float X, float Y);
+public readonly record struct Position(double X, double Y);
